requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
scikit-learn==1.5.0
requests==2.31.0
tensorflow-cpu==2.19.0
